+++ src-tauri/migrations/2024-04-11-112229_add_integ_extra_param/down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE settings
+  DROP COLUMN integration_extra_param;