+++ src-tauri/migrations/2024-04-11-112229_add_integ_extra_param/up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE settings
+  ADD COLUMN integration_extra_param VARCHAR DEFAULT NULL;