+++ src-tauri/src/integrations/redmine.rs
@@ -3,8 +3,26 @@ use crate::models::{GroupedTask, Setting};
 use crate::utils::dates::format_duration;
 use oxhttp::model::{Method, Request, Status};
 use oxhttp::Client;
-use serde_json::json;
-use url::Url;
+use serde::{Deserialize, Serialize};
+use serde_json;
+
+#[derive(Debug, Deserialize)]
+pub struct RedmineError {
+    pub errors: Vec<String>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct RedmineTimeActivity {
+    pub id: i32,
+    pub name: String,
+    pub is_default: bool,
+    pub active: bool,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct RedmineTimeActivities {
+    pub time_entry_activities: Vec<RedmineTimeActivity>,
+}
 
 #[derive(Debug)]
 pub struct Redmine {}
@@ -16,25 +34,26 @@ impl Default for Redmine {
 }
 
 impl Integration for Redmine {
-    fn send_task(&self, settings: &Setting, task: &GroupedTask) -> Result<(), Error> {
-        let mut url = Url::parse(settings.integration_url.as_ref().unwrap()).unwrap();
-        if !url.path().ends_with('/') {
-            url.path_segments_mut().unwrap().push("");
-        }
-        url.path_segments_mut().unwrap().push("time_entries.json");
-
+    fn send_task(
+        &self,
+        settings: &Setting,
+        task: &GroupedTask,
+        extra_param: Option<String>,
+    ) -> Result<(), Error> {
+        let url = Self::prepare_url(settings, vec!["time_entries.json".to_string()]);
         let token = &settings.integration_token.as_ref().unwrap();
-        let body = json!({
+        let body = serde_json::json!({
             "time_entry": {
                 "issue_id": task.external_id,
                 "hours": format_duration(task.duration),
                 "comments": task.desc,
-                "spent_on": task.date.to_string()
+                "spent_on": task.date.to_string(),
+                "activity_id": extra_param,
             }
         });
 
         let client = Client::new();
-        match client.request(Self::prepare_request(
+        match client.request(Self::prepare_post_request(
             url.as_ref(),
             &body.to_string(),
             token,
@@ -46,7 +65,9 @@ impl Integration for Redmine {
                 if response.status() == Status::UNAUTHORIZED {
                     return Err(Error::UnauthorizedError);
                 }
-                Err(Error::CheckExternalIdError)
+                let response_body = response.into_body().to_string().unwrap();
+                let error = serde_json::from_str::<RedmineError>(&response_body).unwrap();
+                Err(Error::GenericError(error.errors))
             }
             Err(_) => Err(Error::UnkownHostError),
         }
@@ -58,7 +79,7 @@ impl Redmine {
         Self {}
     }
 
-    fn prepare_request(url: &str, body: &str, token: &str) -> Request {
+    fn prepare_post_request(url: &str, body: &str, token: &str) -> Request {
         let mut request =
             Request::builder(Method::POST, url.parse().unwrap()).with_body(body.to_string());
         request
@@ -66,5 +87,44 @@ impl Redmine {
             .unwrap();
         request.append_header("X-Redmine-API-Key", token).unwrap();
         request
+    }
+
+    fn prepare_get_request(url: &str, token: &str) -> Request {
+        let mut request = Request::builder(Method::GET, url.parse().unwrap()).build();
+        request
+            .append_header("Content-Type", "application/json")
+            .unwrap();
+        request.append_header("X-Redmine-API-Key", token).unwrap();
+        request
+    }
+
+    pub fn activities(settings: &Setting) -> Vec<RedmineTimeActivity> {
+        let url = Self::prepare_url(
+            settings,
+            vec![
+                "enumerations".to_string(),
+                "time_entry_activities.json".to_string(),
+            ],
+        );
+        let token = &settings.integration_token.as_ref().unwrap();
+        let client = Client::new();
+        match client.request(Self::prepare_get_request(url.as_ref(), token)) {
+            Ok(response) => {
+                if response.status() == Status::OK {
+                    let response_body = response.into_body().to_string().unwrap();
+
+                    let activities =
+                        serde_json::from_str::<RedmineTimeActivities>(&response_body).unwrap();
+
+                    return activities
+                        .time_entry_activities
+                        .into_iter()
+                        .filter(|activity| activity.active)
+                        .collect::<Vec<RedmineTimeActivity>>();
+                }
+                vec![]
+            }
+            Err(_) => vec![],
+        }
     }
 }