+++ src-tauri/src/main.rs
@@ -122,14 +122,14 @@ fn group_tasks() -> Value {
 }
 
 #[command]
-async fn send_to_integration(id: String) -> Result<(), String> {
+async fn send_to_integration(id: String, extra_param: Option<String>) -> Result<(), String> {
     let mut db = db::establish_connection();
     let settings = SettingsRepository::get_settings(&mut db).unwrap();
 
     if let Some(integration) = get_integration(&settings) {
         let tasks = TasksRepository::grouped_tasks(&mut db).unwrap();
         let task = tasks.iter().find(|task| task.id == id).unwrap();
-        match integration.send_task(&settings, task) {
+        match integration.send_task(&settings, task, extra_param) {
             Ok(_) => {
                 let _ = TasksRepository::mark_tasks_as_reported(&mut db, &task.ids.0);
                 Ok(())
@@ -174,6 +174,13 @@ fn favourites() -> Value {
     json!(tasks)
 }
 
+#[command]
+fn redmine_activities() -> Value {
+    let mut db = db::establish_connection();
+    let settings = SettingsRepository::get_settings(&mut db).unwrap();
+    json!(integrations::redmine::Redmine::activities(&settings))
+}
+
 fn main() {
     let system_tray = SystemTray::new();
     tauri::Builder::default()
@@ -196,6 +203,7 @@ fn main() {
             dates_with_tasks,
             toggle_favourite,
             favourites,
+            redmine_activities
         ])
         .system_tray(system_tray)
         .run(tauri::generate_context!())